requirements.txt
flask==3.0.2
requests==2.31.0
PyJWT==2.8.0